BE/requirements.txt
fastapi
uvicorn[standard]
supabase
svix
PyJWT[crypto]
httpx
pydantic-settings
python-dotenv
